+++ src/main/java/gov/usgs/owi/nldi/controllers/LinkedDataControllerV2.java
@@ -51,12 +51,7 @@ public class LinkedDataControllerV2 extends BaseController {
 		HttpServletRequest request, HttpServletResponse response,
 		@PathVariable(LookupDao.FEATURE_SOURCE) String featureSource,
 		@PathVariable(Parameters.FEATURE_ID) String featureID,
-		@PathVariable(Parameters.NAVIGATION_MODE) @Pattern(regexp=REGEX_NAVIGATION_MODE) String navigationMode,
-		@RequestParam(value=Parameters.STOP_COMID, required=false) @Range(min=1, max=Integer.MAX_VALUE) String stopComid,
-		@Parameter(description=Parameters.DISTANCE_DESCRIPTION)
-		@RequestParam(value=Parameters.DISTANCE, required=false, defaultValue=Parameters.MAX_DISTANCE)
-		@Pattern(message=Parameters.DISTANCE_VALIDATION_MESSAGE, regexp=Parameters.DISTANCE_VALIDATION_REGEX) String distance,
-		@RequestParam(value=Parameters.LEGACY, required=false) String legacy) throws Exception {
+		@PathVariable(Parameters.NAVIGATION_MODE) @Pattern(regexp=REGEX_NAVIGATION_MODE) String navigationMode)  {
 
 		BigInteger logId = logService.logRequest(request);