+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 *   Migrate from springfox to springdoc
 *   Added swagger annotations
 *   Implement HTML media type for NLDI resources
+*   Add new API to query for a feature by latitude and longitude
 *   Add V2 for navigate endpoint
   
 ## [1.2.0](https://github.com/ACWI-SSWD/nldi-services/compare/nldi-services-1.1.0...nldi-services-1.2.0)